feat(node): Relocate NodeDetail screen to feature module
This commit refactors the Node Detail screen and its related components, moving them from the `app` module into a new, dedicated `feature/node` module.

Key changes include:
- Moving composables like `NodeDetailContent`, `PositionSection`, `MetricsSection`, `DeviceActions`, and `AdministrationSection` into the `feature/node` module.
- Relocating associated model classes such as `MetricsState`, `TimeFrame`, `NodeDetailAction`, and `LogsType`.
- Migrating drawable resources used by these components.
- Updating the `app` module to use the components from the new `feature/node` module.
- Adding necessary dependencies like maps, coil, and markdown to the `feature/node` module's build file.
