+++ app/src/main/java/com/geeksville/mesh/model/MetricsViewModel.kt
@@ -19,21 +19,15 @@ package com.geeksville.mesh.model
 
 import android.app.Application
 import android.net.Uri
-import androidx.annotation.StringRes
-import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.dp
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import androidx.navigation.toRoute
-import com.geeksville.mesh.ConfigProtos.Config.DisplayConfig.DisplayUnits
 import com.geeksville.mesh.CoroutineDispatchers
 import com.geeksville.mesh.MeshProtos
 import com.geeksville.mesh.MeshProtos.MeshPacket
-import com.geeksville.mesh.MeshProtos.Position
 import com.geeksville.mesh.Portnums
 import com.geeksville.mesh.Portnums.PortNum
-import com.geeksville.mesh.TelemetryProtos.Telemetry
 import com.geeksville.mesh.util.safeNumber
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -55,11 +49,9 @@ import org.meshtastic.core.data.repository.FirmwareReleaseRepository
 import org.meshtastic.core.data.repository.MeshLogRepository
 import org.meshtastic.core.data.repository.NodeRepository
 import org.meshtastic.core.data.repository.RadioConfigRepository
-import org.meshtastic.core.database.entity.FirmwareRelease
 import org.meshtastic.core.database.entity.MeshLog
 import org.meshtastic.core.database.model.Node
 import org.meshtastic.core.model.DataPacket
-import org.meshtastic.core.model.DeviceHardware
 import org.meshtastic.core.navigation.NodesRoutes
 import org.meshtastic.core.prefs.map.MapPrefs
 import org.meshtastic.core.proto.toPosition
@@ -67,126 +59,17 @@ import org.meshtastic.core.service.ServiceAction
 import org.meshtastic.core.service.ServiceRepository
 import org.meshtastic.core.strings.R
 import org.meshtastic.feature.map.model.CustomTileSource
+import org.meshtastic.feature.node.model.MetricsState
+import org.meshtastic.feature.node.model.TimeFrame
 import timber.log.Timber
 import java.io.BufferedWriter
 import java.io.FileNotFoundException
 import java.io.FileWriter
 import java.text.SimpleDateFormat
 import java.util.Locale
-import java.util.concurrent.TimeUnit
 import javax.inject.Inject
 
 private const val DEFAULT_ID_SUFFIX_LENGTH = 4
-
-data class MetricsState(
-    val isLocal: Boolean = false,
-    val isManaged: Boolean = true,
-    val isFahrenheit: Boolean = false,
-    val displayUnits: DisplayUnits = DisplayUnits.METRIC,
-    val node: Node? = null,
-    val deviceMetrics: List<Telemetry> = emptyList(),
-    val signalMetrics: List<MeshPacket> = emptyList(),
-    val powerMetrics: List<Telemetry> = emptyList(),
-    val hostMetrics: List<Telemetry> = emptyList(),
-    val tracerouteRequests: List<MeshLog> = emptyList(),
-    val tracerouteResults: List<MeshLog> = emptyList(),
-    val positionLogs: List<Position> = emptyList(),
-    val deviceHardware: DeviceHardware? = null,
-    val isLocalDevice: Boolean = false,
-    val firmwareEdition: MeshProtos.FirmwareEdition? = null,
-    val latestStableFirmware: FirmwareRelease = FirmwareRelease(),
-    val latestAlphaFirmware: FirmwareRelease = FirmwareRelease(),
-    val paxMetrics: List<MeshLog> = emptyList(),
-) {
-    fun hasDeviceMetrics() = deviceMetrics.isNotEmpty()
-
-    fun hasSignalMetrics() = signalMetrics.isNotEmpty()
-
-    fun hasPowerMetrics() = powerMetrics.isNotEmpty()
-
-    fun hasTracerouteLogs() = tracerouteRequests.isNotEmpty()
-
-    fun hasPositionLogs() = positionLogs.isNotEmpty()
-
-    fun hasHostMetrics() = hostMetrics.isNotEmpty()
-
-    fun hasPaxMetrics() = paxMetrics.isNotEmpty()
-
-    fun deviceMetricsFiltered(timeFrame: TimeFrame): List<Telemetry> {
-        val oldestTime = timeFrame.calculateOldestTime()
-        return deviceMetrics.filter { it.time >= oldestTime }
-    }
-
-    fun signalMetricsFiltered(timeFrame: TimeFrame): List<MeshPacket> {
-        val oldestTime = timeFrame.calculateOldestTime()
-        return signalMetrics.filter { it.rxTime >= oldestTime }
-    }
-
-    fun powerMetricsFiltered(timeFrame: TimeFrame): List<Telemetry> {
-        val oldestTime = timeFrame.calculateOldestTime()
-        return powerMetrics.filter { it.time >= oldestTime }
-    }
-
-    companion object {
-        val Empty = MetricsState()
-    }
-}
-
-/** Supported time frames used to display data. */
-@Suppress("MagicNumber")
-enum class TimeFrame(val seconds: Long, @StringRes val strRes: Int) {
-    TWENTY_FOUR_HOURS(TimeUnit.DAYS.toSeconds(1), R.string.twenty_four_hours),
-    FORTY_EIGHT_HOURS(TimeUnit.DAYS.toSeconds(2), R.string.forty_eight_hours),
-    ONE_WEEK(TimeUnit.DAYS.toSeconds(7), R.string.one_week),
-    TWO_WEEKS(TimeUnit.DAYS.toSeconds(14), R.string.two_weeks),
-    FOUR_WEEKS(TimeUnit.DAYS.toSeconds(28), R.string.four_weeks),
-    MAX(0L, R.string.max),
-    ;
-
-    fun calculateOldestTime(): Long = if (this == MAX) {
-        MAX.seconds
-    } else {
-        System.currentTimeMillis() / 1000 - this.seconds
-    }
-
-    /**
-     * The time interval to draw the vertical lines representing time on the x-axis.
-     *
-     * @return seconds epoch seconds
-     */
-    fun lineInterval(): Long = when (this.ordinal) {
-        TWENTY_FOUR_HOURS.ordinal -> TimeUnit.HOURS.toSeconds(6)
-
-        FORTY_EIGHT_HOURS.ordinal -> TimeUnit.HOURS.toSeconds(12)
-
-        ONE_WEEK.ordinal,
-        TWO_WEEKS.ordinal,
-        -> TimeUnit.DAYS.toSeconds(1)
-
-        else -> TimeUnit.DAYS.toSeconds(7)
-    }
-
-    /** Used to detect a significant time separation between [Telemetry]s. */
-    fun timeThreshold(): Long = when (this.ordinal) {
-        TWENTY_FOUR_HOURS.ordinal -> TimeUnit.HOURS.toSeconds(6)
-
-        FORTY_EIGHT_HOURS.ordinal -> TimeUnit.HOURS.toSeconds(12)
-
-        else -> TimeUnit.DAYS.toSeconds(1)
-    }
-
-    /**
-     * Calculates the needed [Dp] depending on the amount of time being plotted.
-     *
-     * @param time in seconds
-     */
-    fun dp(screenWidth: Int, time: Long): Dp {
-        val timePerScreen = this.lineInterval()
-        val multiplier = time / timePerScreen
-        val dp = (screenWidth * multiplier).toInt().dp
-        return dp.takeIf { it != 0.dp } ?: screenWidth.dp
-    }
-}
 
 private fun MeshPacket.hasValidSignal(): Boolean =
     rxTime > 0 && (rxSnr != 0f && rxRssi != 0) && (hopStart > 0 && hopStart - hopLimit == 0)
