+++ feature/node/src/main/kotlin/org/meshtastic/feature/node/component/EnvironmentMetrics.kt
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.geeksville.mesh.ui.node.components
+package org.meshtastic.feature.node.component
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.FlowRow
@@ -37,14 +37,14 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import com.geeksville.mesh.ConfigProtos
-import com.geeksville.mesh.ui.node.model.DrawableMetricInfo
-import com.geeksville.mesh.ui.node.model.VectorMetricInfo
 import org.meshtastic.core.database.model.Node
 import org.meshtastic.core.model.util.UnitConversions
 import org.meshtastic.core.model.util.UnitConversions.toTempString
 import org.meshtastic.core.model.util.toSmallDistanceString
 import org.meshtastic.core.model.util.toSpeedString
 import org.meshtastic.core.strings.R
+import org.meshtastic.feature.node.model.DrawableMetricInfo
+import org.meshtastic.feature.node.model.VectorMetricInfo
 
 @Suppress("CyclomaticComplexMethod", "LongMethod")
 @Composable
@@ -141,7 +141,7 @@ internal fun EnvironmentMetrics(
                             DrawableMetricInfo(
                                 R.string.dew_point,
                                 dewPoint.toTempString(isFahrenheit),
-                                com.geeksville.mesh.R.drawable.ic_outlined_dew_point_24,
+                                org.meshtastic.feature.node.R.drawable.ic_outlined_dew_point_24,
                             ),
                         )
                     }
@@ -150,7 +150,7 @@ internal fun EnvironmentMetrics(
                             DrawableMetricInfo(
                                 R.string.soil_temperature,
                                 soilTemperature.toTempString(isFahrenheit),
-                                com.geeksville.mesh.R.drawable.soil_temperature,
+                                org.meshtastic.feature.node.R.drawable.soil_temperature,
                             ),
                         )
                     }
@@ -159,7 +159,7 @@ internal fun EnvironmentMetrics(
                             DrawableMetricInfo(
                                 R.string.soil_moisture,
                                 "%d%%".format(soilMoisture),
-                                com.geeksville.mesh.R.drawable.soil_moisture,
+                                org.meshtastic.feature.node.R.drawable.soil_moisture,
                             ),
                         )
                     }
@@ -168,7 +168,7 @@ internal fun EnvironmentMetrics(
                             DrawableMetricInfo(
                                 R.string.radiation,
                                 "%.1f µR/h".format(radiation),
-                                com.geeksville.mesh.R.drawable.ic_filled_radioactive_24,
+                                org.meshtastic.feature.node.R.drawable.ic_filled_radioactive_24,
                             ),
                         )
                     }
