+++ feature/node/src/main/res/drawable/hw_unknown.xml
