+++ feature/node/src/main/res/drawable/ic_filled_radioactive_24.xml
