+++ feature/node/src/main/res/drawable/ic_outlined_dew_point_24.xml
