+++ feature/node/src/main/res/drawable/soil_moisture.xml
