+++ feature/node/src/main/res/drawable/soil_temperature.xml
